Admin area

+++ PetHotel/Areas/Client/Controllers/PetController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.Extensions.Caching.Memory;
+using PetHotel.Common;
 using PetHotel.Core.Contracts;
 using PetHotel.Core.Models.PetModels;
 using System.Security.Claims;
@@ -14,10 +15,12 @@ namespace PetHotel.Areas.Client.Controllers
     {
         private readonly IPetService service;
         private readonly IMemoryCache cache;
-        public PetController(IPetService _service, IMemoryCache _cache)
+        private readonly IUserService userService;
+        public PetController(IPetService _service, IMemoryCache _cache, IUserService _userService)
         {
             this.service = _service;
             this.cache = _cache;
+            this.userService = _userService;
         }
         public IActionResult Index()
         {
@@ -28,10 +31,9 @@ namespace PetHotel.Areas.Client.Controllers
 
         public async Task<IActionResult> Create()
         {
-
+            if (!User.IsInRole(GlobalConstants.UserRoleName)) return RedirectToAction("Index", "Home");
             ViewData["PetTypeId"] = new SelectList(await GetTypes(), "Id", "Name");
 
-           /* cache.Set("PetTypeId", types);*/
             return View();
         }
 
@@ -57,32 +59,30 @@ namespace PetHotel.Areas.Client.Controllers
 
         public async Task<IActionResult>Delete(int Id)
         {
-            try
-            {
-                await service.DeletePetAsync(Id);
-            }
-            catch (Exception)
-            {
+            if (!User.IsInRole(GlobalConstants.UserRoleName)) return RedirectToAction("Index", "Home");
 
-                return BadRequest();
-            }
+            string userId = User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.NameIdentifier)!.Value;
+
+            if (await userService.UserOwnsPet(userId, Id)) await service.DeletePetAsync(Id);
+
+           
             return RedirectToAction("Profile", "User");
         }
         [Route("Edit")]
         [HttpGet]
         public async Task<IActionResult> Edit(int Id)
         {
-            try
-            {
-                var model = await service.GetPetAsync(Id);
-                ViewData["PetTypeId"] = new SelectList(await GetTypes(), "Id", "Name");
-                return View(model);
-            }
-            //todo
-            catch (ArgumentNullException)
-            {
-               return BadRequest();
-            }
+            if (!User.IsInRole(GlobalConstants.UserRoleName)) return RedirectToAction("Index", "Home");
+
+            string userId = User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.NameIdentifier)!.Value;
+
+            if (!await userService.UserOwnsPet(userId, Id)) return RedirectToAction("Profile", "User"); 
+
+
+            var model = await service.GetPetAsync(Id);
+            ViewData["PetTypeId"] = new SelectList(await GetTypes(), "Id", "Name");
+           return View(model);
+            
             
         }
 
@@ -95,16 +95,10 @@ namespace PetHotel.Areas.Client.Controllers
                 ViewData["PetTypeId"] = new SelectList(await GetTypes(), "Id", "Name");
                 return View(model); 
             }
-            try
-            {
+            
                 await service.EditPetAsync(model);
                 return RedirectToAction("Profile", "User");
-            }
-            catch (ArgumentNullException)
-            {
-
-                return BadRequest();
-            }
+            
            
         }
